+++ src/main/java/org/yinwang/pysonar/ast/Comprehension.java
@@ -9,12 +9,19 @@ public class Comprehension extends Node {
     public Node target;
     public Node iter;
     public List<Node> ifs;
+    public boolean isAsync;
 
     public Comprehension(Node target, Node iter, List<Node> ifs, String file, int start, int end, int line, int col) {
+        this(target, iter, ifs, false, file, start, end, line, col);
+    }
+
+    public Comprehension(Node target, Node iter, List<Node> ifs, boolean isAsync,
+        String file, int start, int end, int line, int col) {
         super(NodeType.COMPREHENSION, file, start, end, line, col);
         this.target = target;
         this.iter = iter;
         this.ifs = ifs;
+        this.isAsync = isAsync;
         addChildren(target, iter);
         addChildren(ifs);
     }