+++ src/main/java/org/yinwang/pysonar/ast/FunctionDef.java
@@ -17,6 +17,7 @@ public class FunctionDef extends Node {
     public final List<Node> kwOnlyArgs;
     public final List<Node> kwDefaults;
     public final List<Node> annotations;
+    public final List<TypeParameter> typeParams;
     @Nullable
     public final Node returnAnnotation;
     public final int posOnlyArgCount;
@@ -30,6 +31,7 @@ public class FunctionDef extends Node {
         Name vararg, Name kwarg, List<Node> decorators, String file, boolean isAsync, int start, int end, int line, int col) {
         this(name, args, body, defaults, vararg, kwarg, decorators,
                 new ArrayList<>(), new ArrayList<>(), new ArrayList<>(), null, 0,
+                new ArrayList<>(),
                 file, isAsync, start, end, line, col);
     }
 
@@ -37,6 +39,7 @@ public class FunctionDef extends Node {
         Name vararg, Name kwarg, List<Node> decorators,
         List<Node> kwOnlyArgs, List<Node> kwDefaults, List<Node> annotations,
         @Nullable Node returnAnnotation, int posOnlyArgCount,
+        List<TypeParameter> typeParams,
         String file, boolean isAsync, int start, int end, int line, int col) {
         super(NodeType.FUNCTIONDEF, file, start, end, line, col);
         if (name != null) {
@@ -59,6 +62,7 @@ public class FunctionDef extends Node {
         this.annotations = annotations;
         this.returnAnnotation = returnAnnotation;
         this.posOnlyArgCount = posOnlyArgCount;
+        this.typeParams = typeParams;
         this.isAsync = isAsync;
         addChildren(name);
         addChildren(args);
@@ -68,6 +72,7 @@ public class FunctionDef extends Node {
         addChildren(decorators);
         addChildren(annotations);
         addChildren(returnAnnotation);
+        addChildren(typeParams);
         addChildren(vararg, kwarg, this.body);
     }